requirements.txt
asgiref==3.8.1
bleach==6.2.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.3
django-crispy-forms==1.9.1
django-import-export==4.3.7
django-summernote==0.8.20.0
et_xmlfile==2.0.0
MarkupPy==1.18
odfpy==1.4.1
openpyxl==3.1.5
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
webencodings==0.5.1
xlrd==2.0.2
xlwt==1.3.0
